Update build scripts for Linux ARM64
This doesn't include an ARM updater, so those lines in build.sh need to
be commented out for the build to succeed.

Addresses #3515

+++ app/update-packaging/build_autoupdate.sh
@@ -140,15 +140,16 @@ for version in "$FROM" "$TO"; do
 	WIN32_ARCHIVE="Zotero-${version}_win32.zip"
 	WIN_X64_ARCHIVE="Zotero-${version}_win-x64.zip"
 	WIN_ARM64_ARCHIVE="Zotero-${version}_win-arm64.zip"
-	LINUX_X86_ARCHIVE="Zotero-${version}_linux-i686.tar.bz2"
-	LINUX_X86_64_ARCHIVE="Zotero-${version}_linux-x86_64.tar.bz2"
+	LINUX_X86_ARCHIVE="Zotero-${version}_linux-i686.tar.xz"
+	LINUX_X86_64_ARCHIVE="Zotero-${version}_linux-x86_64.tar.xz"
+	LINUX_ARM64_ARCHIVE="Zotero-${version}_linux-arm64.tar.xz"
 	
 	CACHE_DIR="$ROOT_DIR/cache"
 	if [ ! -e "$CACHE_DIR" ]; then
 		mkdir "$CACHE_DIR"
 	fi
 	
-	for archive in "$MAC_ARCHIVE" "$WIN32_ARCHIVE" "$WIN_X64_ARCHIVE" "$WIN_ARM64_ARCHIVE" "$LINUX_X86_ARCHIVE" "$LINUX_X86_64_ARCHIVE"; do
+	for archive in "$MAC_ARCHIVE" "$WIN32_ARCHIVE" "$WIN_X64_ARCHIVE" "$WIN_ARM64_ARCHIVE" "$LINUX_X86_ARCHIVE" "$LINUX_X86_64_ARCHIVE" "$LINUX_ARM64_ARCHIVE"; do
 		if [[ $archive = "$MAC_ARCHIVE" ]] && [[ $BUILD_MAC != 1 ]]; then
 			continue
 		fi
@@ -165,6 +166,9 @@ for version in "$FROM" "$TO"; do
 			continue
 		fi
 		if [[ $archive = "$LINUX_X86_64_ARCHIVE" ]] && [[ $BUILD_LINUX != 1 ]]; then
+			continue
+		fi
+		if [[ $archive = "$LINUX_ARM64_ARCHIVE" ]] && [[ $BUILD_LINUX != 1 ]]; then
 			continue
 		fi
 		
@@ -244,14 +248,14 @@ for version in "$FROM" "$TO"; do
 	
 	# Unpack Linux tarballs
 	if [ $BUILD_LINUX == 1 ]; then
-		if [[ -f "$LINUX_X86_ARCHIVE" ]] && [[ -f "$LINUX_X86_64_ARCHIVE" ]]; then
-			for build in "$LINUX_X86_ARCHIVE" "$LINUX_X86_64_ARCHIVE"; do
-				tar -xjf "$build"
+		if [[ -f "$LINUX_X86_ARCHIVE" ]] && [[ -f "$LINUX_X86_64_ARCHIVE" ]] && [[ -f "$LINUX_ARM64_ARCHIVE" ]]; then
+			for build in "$LINUX_X86_ARCHIVE" "$LINUX_X86_64_ARCHIVE" "$LINUX_ARM64_ARCHIVE"; do
+				tar -xf "$build"
 				rm "$build"
 			done
 			INCREMENTALS_FOUND=1
 		else
-			echo "$LINUX_X86_ARCHIVE/$LINUX_X86_64_ARCHIVE not found"
+			echo "$LINUX_X86_ARCHIVE/$LINUX_X86_64_ARCHIVE/$LINUX_ARM64_ARCHIVE not found"
 		fi
 	fi
 	
@@ -263,7 +267,7 @@ export MOZ_PRODUCT_VERSION="$TO"
 export MAR_CHANNEL_ID="$CHANNEL"
 
 CHANGES_MADE=0
-for build in "mac" "win32" "win-x64" "win-arm64" "linux-i686" "linux-x86_64"; do
+for build in "mac" "win32" "win-x64" "win-arm64" "linux-i686" "linux-x86_64" "linux-arm64"; do
 	if [[ $build == "mac" ]]; then
 		if [[ $BUILD_MAC == 0 ]]; then
 			continue
@@ -273,7 +277,7 @@ for build in "mac" "win32" "win-x64" "win-arm64" "linux-i686" "linux-x86_64"; do
 		if ([[ $build == "win32" ]] || [[ $build == "win-x64" ]] || [[ $build == "win-arm64" ]]) && [[ $BUILD_WIN == 0 ]]; then
 			continue
 		fi
-		if ([[ $build == "linux-i686" ]] || [[ $build == "linux-x86_64" ]]) && [[ $BUILD_LINUX == 0 ]]; then
+		if ([[ $build == "linux-i686" ]] || [[ $build == "linux-x86_64" ]] || [[ $build == "linux-arm64" ]]) && [[ $BUILD_LINUX == 0 ]]; then
 			continue
 		fi
 		dir="Zotero_$build"
